[win32] network: move implementation to platform folder.
