+++ cmake/treedata/windows/subdirs.txt
@@ -1,10 +1,10 @@
 xbmc/platform/win32             platform/win32
 xbmc/platform/win32/input       platform/win32/input
 xbmc/platform/win32/filesystem  platform/win32/filesystem
+xbmc/platform/win32/network     platform/win32/network
 xbmc/platform/win32/storage     platform/win32/storage
 xbmc/input/touch                input/touch
 xbmc/input/touch/generic        input/touch/generic
-xbmc/network/windows            network/windows
 xbmc/network/mdns               network/mdns
 xbmc/peripherals/bus/win32      peripherals/bus/win32
 xbmc/powermanagement/windows    powermanagement/windows
