+++ xbmc/network/Network.h
@@ -157,7 +157,7 @@ public:
 #elif defined(HAS_LINUX_NETWORK)
 #include "linux/NetworkLinux.h"
 #elif defined(HAS_WIN32_NETWORK)
-#include "windows/NetworkWin32.h"
+#include "platform/win32/network/NetworkWin32.h"
 #elif defined(HAS_WIN10_NETWORK)
 #include "platform/win10/network/NetworkWin10.h"
 #endif
