+++ xbmc/platform/win32/network/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES NetworkWin32.cpp)
 
 set(HEADERS NetworkWin32.h)
 
-core_add_library(network_windows)
+core_add_library(platform_win32_network)
