+++ xbmc/platform/win32/network/NetworkWin32.h
